requirements.txt
asgiref==3.8.1
bleach==6.1.0
crispy-bootstrap4==2023.1
Django==5.0.6
django-crispy-forms==2.1
django-filter==23.2
django-resized==1.0.2
django-summernote==0.8.20.0
Pillow==10.0.0
python-decouple==3.6
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
webencodings==0.5.1
